+++ src/Screen/Actions/DropDown.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Orchid\Screen\Actions;
 
-use Orchid\Screen\Action;
-use Orchid\Screen\Contracts\ActionContract;
-use Orchid\Screen\Repository;
+use Orchid\Screen\Action;use Orchid\Screen\Contracts\Actionable;use Orchid\Screen\Repository;
 
 /**
  * Class DropDown.
@@ -52,7 +50,7 @@ class DropDown extends Action
     }
 
     /**
-     * @param ActionContract[] $list
+     * @param Actionable[] $list
      *
      * @return DropDown
      */